Removed Dashboard packages (codebase split in 2 repos)
