--- components/dashboard/layouts/layoutListItem/LayoutListItem.module.css
@@ -1,27 +1,0 @@
-.container {
-  background-color: var(--white);
-  border-top: 1px solid var(--gray-medium);
-  padding: 15px;
-  /* border-radius: var(--radius-medium); */
-}
-
-.container:hover {
-  background-color: var(--gray-superlight);
-}
-
-.container:first-child {
-  border-top: none;
-}
-
-.title {
-  color: var(--black);
-  font-size: 18px;
-  font-weight: var(--font-medium);
-}
-
-.variations {
-  font-size: 12px;
-  color: var(--gray-strong);
-  font-weight: var(--font-regular);
-  margin-top: 5px;
-}
