--- components/dashboard/sidebar/button/SidebarButton.module.css
@@ -1,33 +1,0 @@
-.button {
-  color: var(--gray-dark);
-  font-weight: var(--font-medium);
-  text-transform: uppercase;
-  background-color: var(--gray-superlight);
-  border: 2px solid var(--gray-light);
-  border-radius: var(--radius-medium);
-  padding: 6px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: left;
-  gap: 10px;
-  font-size: 15px;
-  transition: var(--transition);
-}
-
-.button:hover {
-  /* box-shadow: var(--effect-shadow); */
-  background-color: var(--gray-light);
-  border-color: var(--gray-medium);
-}
-
-.button i {
-  font-size: 25px;
-}
-
-.button.active {
-  color: var(--white);
-  background-color: var(--theme-fill);
-  border-color: var(--theme-border);
-  box-shadow: var(--effect-shadow);
-}
