--- components/dashboard/sidebar/container/Sidebar.module.css
@@ -1,44 +1,0 @@
-.container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  background-color: var(--white);
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: stretch;
-  justify-content: stretch;
-}
-
-.sidebar {
-  width: 250px;
-  min-width: 250px;
-  max-width: 250px;
-  background-color: var(--white);
-  padding: 30px;
-  box-shadow: var(--effect-shadow);
-  z-index: 999;
-}
-
-.buttonsContainer {
-  display: grid;
-  gap: 15px;
-}
-
-.pageContainer {
-  width: 100%;
-  padding: 50px;
-  overflow: scroll;
-  background-color: var(--gray-light);
-}
-
-.pageTitle {
-  font-size: 30px;
-  font-weight: var(--font-bold);
-  margin-bottom: 30px;
-}
-
-.pageContent {
-  width: auto;
-}
