+++ package.json
@@ -11,7 +11,7 @@
     "classnames": "^2.3.1",
     "lodash": "^4.17.21",
     "lottie-web": "^5.9.1",
-    "mongoose": "^6.2.4",
+    "mongoose": "^6.2.6",
     "next": "^12.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
@@ -19,7 +19,7 @@
   },
   "devDependencies": {
     "@svgr/webpack": "^6.2.1",
-    "eslint": "8.10.0",
+    "eslint": "8.11.0",
     "eslint-config-next": "^12.1.0",
     "typescript": "^4.6.2"
   }
