--- pages/dashboard/customers/index.js
@@ -1,5 +1,0 @@
-import Sidebar from '../../../components/dashboard/sidebar/container/Sidebar';
-
-export default function DashboardCustomers() {
-  return <Sidebar>Customers</Sidebar>;
-}
